Fixing Issue #353 Capture config users&roles by name
The setup:get-configuration function in setup.xqy used only ids to
return users and roles. This prevented the capture configurations
option from capturing users and roles specified by name in the
current project configuration or as a parameter. This was a
documented “TODO” in the code in order to improve the capture
functionality. The setup:get-configuration function will now take
user/role names or ids as a parameter and return the requested
configurations. Also added default users and roles from current
project to capture_environment_config in server_config.rb. Lastly
removed word “full” from logging since it could be a full or partial
configuration that is being captured.