Fixing Issue #223 Capture modules for any app type
Currently Capture was only supporting capturing modules for
REST applications. I have added the ability to capture modules
for any application type to ML7 and taken the protection off
which will allow to capture modules for any application type in
ML5,6, and 7. For ML7 if it’s a REST application, REST will still be
used to perform the capture since it’s faster than the query
based method. The query based method used for ML5,6, and
ML7 (non-REST) only grabs text based files (xqy, html,…) it does
not correctly capture other types of files in the modules database
such as images since it’s doing an fn:doc on the URI to get the
content.