Fixing Issue #499 Configuration with Large Number of Indexes
With a large number of indexes the bootstrapping process was taking
a significant amount of time, beyond the default Roxy HTTP configuration.
After looking at setup.xqy it was found that some of the index removal
calls were iterating over the items instead of calling the removal function
with the entire set. Similarly, some of the index creation functions were
recursively adding the items instead of calling the admin function with the
entire set. After looking at docs for all supported ML versions it was found
that calling the functions with all items was supported.
Changes were made to remove and add indexes directly. This made a large
impact in the bootstrapping time for scenarios with a large number of
configuration items.
As a reference, in a test setup with around 1,200 configuration items the
bootstrap operation was reduced from around 20 minutes to under
2 minutes with the above being the only change.