Fix #559 Provide an assert-equal-xml helper
This implements an assert-equal-xml method that provides a structured
comparison of two XML nodes. It displays the error on the first
mismatch between the structure of the two trees, with the paths
to the current node being compared.

This makes it easier to locate where two XML documents are different.