Switch the xcc port to https if https is enabled.
When building the url to talk to the xcc port roxy doesn't look whether that should be https or not,
it's just hardcoded to http, which makes it impossible to deploy modules. This commit changes that
by switching to https if either `use-https-for-xcc` is set to true. By default this option is set to
false.